GdkGLContext: Never use glXSwapIntervalSGI on composited WMs
Delaying the copy of the backbuffer to the front buffer until
the vblank doesn't make sense when the frontbuffer is not the
real screen, as the copy will not flicker anyway. It is the
responsibility of the compositor to sync *its* drawing of the
window to the vblank.

The other reason to delay to vblank is to avoid rendering all the
time and sync each frame to the actual frame rate. However, we
already handle this in Gtk+ via the frame clock, so this should
not be needed.