Initial work at using FBOs for gl rendering
This converts the GtkGLArea to do all rendering to an FBO, which is
then rendered to the screen with the new gdk_cairo_draw_gl_framebuffer()
call.

At the moment there is only a slow glReadPixels based version of
gdk_cairo_draw_gl_framebuffer(), but the goal is to be able to do
much better with this.