Accelerate glarea rendering
When we draw the render buffer in gtkglarea we do this directly
via gl to the back buffer, and we also clear the drawn area in the
normal double buffer surface to transparent. This means when we later
draw the double buffer surface to the back buffer the gl content will
be seen (unless the app drew more on the double buffer).

The gl drawing onto the back buffer is hit on a fast-path in
gdk_cairo_draw_gl_render_buffer() that is only enabled if drawing to
the real cairo target (i.e. to the window, with no push_group()
active) with a trivial transformation. If any of this is not right we
fall back to the software path that reads back the render buffer pixels
and composites with cairo.