Remove GdkGLPixelFormat and change how gl contexts are created
The way GdkGLPixelFormat and the framebuffer setup worked matched the
traditional GLX way that GL has worked where you create a native
window and do negotiations for the framebuffer details on that.

This doesn't really match the current model where Gdk owns the
gl rendering and presenting to the toplevel and child widgets work
by using renderbuffers. Additionally, it completely fails on wayland
as there we have to actually allocate the surface buffer for the
toplevel, and there can be only one, with only one framebuffer
config, so we can't expose some API to let the user create different
ones.

So, in the new world we change context creation to:

GdkGLContext *gdk_window_create_gl_context (GdkWindow    *window,
                                            GdkGLProfile  profile,
                                            GError      **error);

Which creates a context for the native window that backs this
particular window, which cannot be bound to another window. The
framebuffer config chosen for the context is the one described by the
window visual (i.e. same bitdept, has alpha if window is rgba visual,
no aux buffers). Anything that requires rendering with other kinds
of buffers (stencil, depth, etc) need to do that with render buffers.

This simplifies the APIs a lot, and allows a sane wayland
implementation. Limiting the context to a single window also fixes
some remaining issues with making contexts current.