Implement buffer age optimization to avoid redrawing a lot
This uses the buffer age extension to detect the "age" of the
current backbuffer, to avoid redrawing things that is already
uptodate in the backbuffer. It also adds a new green color
for the update region to show the area of the window that
was redrawn due to having to repaint old areas of the back buffer.