widget: Add GskLayer integration
Each GtkWidget can be set up to be the entry point of a scene graph, by
adding a GskLayer to the components of the widget itself. This means
that it's possible to set up a GSK-based scene graph when creating a
widget, and the widget will then be in charge of drawing it whenever
it's needed.