gsk: Add GskLayoutManager
The GskLayoutManager type is an interface to be implemented by delegate
objects that want to be responsible for the layout of children of a
specific layer.

GskLayoutManager implementations can provide overrides for the layout
management related virtual functions inside GskLayer; GskLayer itself
will query the layout manager instance, if one is set up.